chore: renaming assets / refac imports
