--- frontend/constants/ProductList.json
@@ -1,39 +1,0 @@
-[
-    {
-        "name": "Dark Roast",
-        "description": "savory",
-        "price": 12.75,
-        "grind": "WholeBean",
-        "image": ""
-    },
-    {
-        "name": "Expresso",
-        "description": "For that extra punch",
-        "price": 12.75,
-        "grind": "WholeBean",  
-        "image": ""
-    },
-    {
-        
-        "name": "Light Roast",
-        "description": "Caramel and vanilla flavoring",
-        "price": 12.75,
-        "grind": "WholeBean",
-        "image": ""
-    },
-    {
-        "name": "Medium Roast",
-        "description": "A happy medium between light and dark",
-        "price": 12.75,
-        "grind": "WholeBean",
-        "image": ""
-    },
-    {
-        "name": "Decaf",
-        "description": "No caffiene ",
-        "price": 12.75,
-        "grind": "WholeBean",
-        "image": ""
-    }
-
-]
